fix: triangle and noise channels sounding wrong

+++ umesen-core/src/cpu/mod.rs
@@ -7,8 +7,8 @@ pub use disassembler::Disassembler;
 pub use opcode::{AddrMode, Inst, Opcode};
 
 /// Number of clock cycles per second
-pub const CLOCK_SPEED_HZ: f64 = 1789773.;
-pub const CYCLES_PER_FRAME: f64 = 29780.5;
+pub const CLOCK_SPEED_HZ: f32 = 1789773.;
+pub const CYCLES_PER_FRAME: f32 = 29780.5;
 
 bitflags::bitflags! {
     /// Flags for the cpu register
